Relax the restrictions on argument types to a unique_ptr instantiated on an array type.
Relax the restrictions on argument types to a unique_ptr
instantiated on an array type.  This patch is a backport
of Jonathan Wakely's "std::unique_ptr<T[], D> improvements"
http://gcc.gnu.org/ml/gcc-patches/2012-12/msg01271.html to the
google 4.7 branch.

The existing unique_ptr admits no conversions, not even
cv-qualification conversions.  The patch permits cv-qualification
conversion.  It prohibits derived-to-base pointer conversion, as
is needed to prevent improper deletion.  It permits user-defined
pointer types.  The latter permissiveness is under debate.  Given

  struct base { };
  struct derived : base { };

as we cannot presently distinguish between the safe

  struct ptr_base {
    operator base*() { return 0; }
  };

  ... unique_ptr< base >( ptr_base() );

and the unsafe

  struct ptr_derived {
    operator derived*() { return 0; }
  };

  ... unique_ptr< base >( ptr_derived() );

No existing code can encounter the unsafe part because no existing
code can have user-defined pointer types.  There are likely to be
very few uses of user-defined pointers introduced into the code
base between now and when the issue is resolved.  So, backporting
the patch as is seems safe enough, and not too much trouble to
correct later.

Tested on x86_64.  Tests pass through Google's core and mantle,
but fail for apparently unrelated reasons in crust.

From-SVN: r194874