This patch adds -g[no-]fission as a synonym for -g[no-]split-dwarf,
and modifies the existing stub we have for -gmlt to make it a synonym
for -g1. The implementation of -gsplit-dwarf is already in trunk (and
therefore g/i), and the implementation of -g1 to produce line tables
will be submitted for google/main shortly, and for trunk when Stage 1
reopens.

From-SVN: r195824