Merge revision 187577 from branch google/gcc-4_7:
Allow static const floats unless -pedantic is passed.

This patch allows us to migrate to C++11 more incrementally, since we can leave
the static const float initializations in place, flip the switch, and then
change it to use constexpr.

This is a forward port of r180638 from google/gcc-4_6 (despite the fact that
that revision says to NOT forward-port this).  Minor additional fixups have
been applied.

2012-05-15   Ollie Wild  <aaw@google.com>

        * gcc/cp/decl.c (check_static_variable_definition): Only generate a
        constexpr warning when -pedantic is enabled.
        * gcc/testsuite/g++.dg/cpp0x/constexpr-static8.C: Replace -fpermissive
        with -pedantic.
        * gcc/testsuite/g++.dg/cpp0x/constexpr-static8_nonpedantic.C: New test.
        * gcc/testsuite/g++.old-deja/g++.ext/memconst.C: Compile with -pedantic
        -pedantic-errors to work around test failures with -std=gnu++11.

From-SVN: r195904