Patch for the reordering plugin to display appropriate error messages when...
Patch for the reordering plugin to display appropriate error messages
when incorrect options are passed or when some API is unavailable.

   The plugin will use the ld_plugin_message linker API to flag errors
(fatal or non-fatal) when incorrect options are passed or when the
required API is not available in the invoking linker. If the message
API is missing, then it falls back to using fprintf.

From-SVN: r196040