Fix the nightly build of gcc-4.7 based ppc32 crosstool and ppc64 crosstool.
What happened was that libmudflap configure looks for a preprocessor by trying $CXX -E and then backing off to /lib/cpp.  $CXX -E is failing with "unrecognized command line option ‘-funconfigured-libstdc++’", and the /lib/cpp backstop then fails also. The -funconfigured-libstdc++ is because configure can't find libstdc++/scripts/testsuite_flags. This is a so-far undiagnosed race in gcc make, masked where /lib/cpp is available.   And that's absent
because in this build, for whatever reason, libstdc++ loses a race
with libmudflap.

2013-03-12  Jing Yu  <jingyu@google.com>

	* Makefile.def (configure-target-libmudflap): add dependence on
	configure-target-libstdc++-v3.
	* Makefile.in: regenerate.

From-SVN: r196618