This patch restores part of r196176...
This patch restores part of r196176, which was subsequently reverted
due to issues with the other part of that patch that dealt with
module info updates. This patch restores the option -femit-function-names
to emit to stderr the mapping from module name/function id to function
assembler name. This enables mapping the function ids in the gcda
dump files to the symbolic function name. This will work for LIPO builds
without the other part of that patch, which was needed for FDO builds,
and which needs to be revisited and fixed.

2013-03-31  Teresa Johnson  <tejohnson@google.com>

	* tree-profile.c (tree_profiling): Optionally call
        new function emit_function_name.
	* common.opt (flag_emit_function_names): New flag.
	* coverage.c (emit_function_name): New function.
	* coverage.h: Ditto.

From-SVN: r197289