The enclosed patch puts strings in the .debug_str section in the .o file with -gsplit-dwarf.
The enclosed patch puts strings in the .debug_str section in the .o
file with -gsplit-dwarf. It is backported from from trunk (which, in turn,
was ported from google_4.7).

http://gcc.gnu.org/ml/gcc-patches/2013-04/msg01355.html

The patch itself is no different from the one finally applied to trunk,
except several hunks have moved a few lines.

From-SVN: r198481