Google ref b/9628617
GCC trunk supports calling x86 intrinsics from select functions in a file that are tagged with the corresponding target attribute without having to compile the entire file with the -mxxx option. This improves the usability of x86 intrinsics. This was ported to google/gcc-4_8 in rev. 200435 and reverted in rev. 200497 because of bug http://gcc.gnu.org/bugzilla/show_bug.cgi?id=57698

This is being ported again with fixes for PR57698 in trunk rev.201039 included along with trunk rev. 200627 back-ported to fix inlining of functions with avx target attribute.

From-SVN: r201046