This patch ports the remaining -fopt-info messages that had been added to google/gcc-4_7 using the original -fopt-info framework implemented on that branch...
This patch ports the remaining -fopt-info messages that had been added
to google/gcc-4_7 using the original -fopt-info framework implemented
on that branch, to google/gcc-4_8 using the new -fopt-info framework
from trunk.

Specifically, this ports over the messages added/modified by r180269,
r180973 and r195968. Some of these patches/messages were already ported,
but this adds the remainder. I converted from the old OPT_INFO_* levels
to the new MSG_* levels using the following mapping:
  OPT_INFO_MIN -> MSG_OPTIMIZED_LOCATIONS
  OPT_INFO_MED -> MSG_MISSED_OPTIMIZATION
  OPT_INFO_MAX -> MSG_NOTE
The affected messages relate to LIPO module imports, missing/mismatched/
corrupted profile data, and indirect call promotions performed.

Additionally, after discussing with Easwaran and Rong, I flipped the
default of the inline-dump-module-id parameter to 1 to enable dumping
module ids in -fopt-info inlining messages for easier inlining
report generation.

Finally, I had to make a couple enhancements to the new dump infrastructure
to get some of the new messages emitted. This change should be
pushed to trunk as it will help facilitate adding new dump messages
more easily (particularly number 2 below). There were two issues:
1) The module imports occurred during coverage_init which happens
very early, before entering the pass manager that sets up the dumping.
This was addressed by enabling dumping within coverage_init, and
is modeled on how dumping is enabled after the pass manager
during finish_optimization_passes.
2) Dumping was only enabled for passes marked with an optinfo_flag
that isn't OPTGROUP_NONE. Currently optgroup flags are only setup for
optimization groups in the categories IPA, LOOP, INLINE and VEC.
What this means is that any dump messages added to a pass that isn't
in one of these groups will silently be suppressed. The OPTGROUP
setting is specified in opt_pass struct, and is also automatically set
to OPTGROUP_IPA for any pass starting with "ipa-". What I did was
to add a new optgroup macro, OPTGROUP_OTHER. This is enabled only
under -fopt-info-optall, which is also the default for -fopt-info.
That way dump messages can be emitted without the requirement that
the pass be part of a group that can be emitted under a specific
optimization group subset. When setting up the dumps, any pass that
has OPTGROUP_NONE after examining the opt_pass struct and the pass
name will use OPTGROUP_OTHER. This doesn't mean that the list of
optgroups shouldn't be expanded, but rather adds a catch-all for
passes that don't currently have or need to be emitted on their own
as part of a new optgroup.

2013-07-25  Teresa Johnson  <tejohnson@google.com>

	* c-family/c-opts.c (lipo_max_mem_reached): Use new dump framework.
	* profile.c (read_profile_edge_counts): Ditto.
	(compute_branch_probabilities): Ditto.
	* value-prof.c (check_counter): Ditto.
	(check_ic_counter): Ditto.
	(find_func_by_funcdef_no): Ditto.
	(check_ic_target): Ditto.
	(gimple_ic_transform_mult_targ): Ditto.
	* mcf.c (find_minimum_cost_flow): Ditto.
	* coverage.c (incompatible_cl_args): Ditto.
	(read_counts_file): Ditto.
	(get_coverage_counts): Ditto.
	(coverage_init): Setup new dump framework.
	* common.opt (flag_ripa_verbose): Remove.
	* params.def (PARAM_INLINE_DUMP_MODULE_ID): Enable by default.
	* dumpfile.h (OPTGROUP_OTHER): Add and enable under OPTGROUP_ALL.
	* passes.c (register_one_dump_file): Use OPTGROUP_OTHER
        when pass not in any opt group.
	* doc/invoke.texi: Document optall -fopt-info flag.

	* testsuite/gcc.dg/pr32773.c: Use -fopt-info.
	* testsuite/gcc.dg/pr40209.c: Ditto.
	* testsuite/gcc.dg/pr26570.c: Ditto.
	* testsuite/gcc.dg/tree-prof/lipo/lipo_inline1_0.c: Ditto.
	* testsuite/g++.dg/tree-ssa/dom-invalid.C: Ditto.

From-SVN: r201269