Handle NULL vdef for call in the case where we have a matching vnresult that...
Handle NULL vdef for call in the case where we have a matching vnresult
that has a vdef (already handle the NULL vdef case when !vnresult). This
can happen for promoted indirect calls if the fallback indirect call
(which has a vdef) can be proven equivalent to the promoted direct call
(which might not have a vdef).

2014-01-14  Teresa Johnson  <tejohnson@google.com>

        Google ref b/12531084.
	* tree-ssa-sccvn.c (visit_reference_op_call): Handle NULL vdef.

Index: tree-ssa-sccvn.c
===================================================================
--- tree-ssa-sccvn.c	(revision 206602)
+++ tree-ssa-sccvn.c	(working copy)
@@ -2725,7 +2725,7 @@ visit_reference_op_call (tree lhs, gimple stmt)
 
   if (vnresult)
     {
-      if (vnresult->result_vdef)
+      if (vnresult->result_vdef && vdef)
 	changed |= set_ssa_val_to (vdef, vnresult->result_vdef);
 
       if (!vnresult->result && lhs)

From-SVN: r206604