Remove check for "Invalid sum" in update-loopch_0.c to avoid failures after trunk merge.
Remove check for "Invalid sum" in update-loopch_0.c to avoid failures
after trunk merge. Trunk now dumps bbs when they are removed when
the "details" dump option is specified, and since the blocks have had
their edges removed already, profile insanities are detected and flagged
in the dump output. This is confusing the test dump scanner which is
looking for profile insanities in the dump output.

From-SVN: r209220