This patch fixes an issue where self-recursive calls were missed during ipa inlining in LIPO compiles...
This patch fixes an issue where self-recursive calls were missed
during ipa inlining in LIPO compiles, since the resolved nodes were
not checked. When a self-recursive call was inlined incorrectly,
ipa_inline redirected the edge to the resolved node, leading to
a cycle in the cgraph that wasn't expected and resulted in an
infinite loop in estimate_calls_size_and_time.

Also fix an inconsistency in the declaration of include_all_aux
that was exposed when l-ipo.h was included in cgraph.

Google ref b/13912450.

2014-04-10  Teresa Johnson  <tejohnson@google.com>

	* cgraph.h (cgraph_edge_recursive_p): Check the resolved
        node in LIPO compiles.
	* l-ipo.h (include_all_aux): Fix declaration that didn't
        match definition.
	* Makefile.in (CGRAPH_H): Include l-ipo.h.

From-SVN: r209284