This patch ensures we don't prematurely delete an abstract origin node...
This patch ensures we don't prematurely delete an abstract origin node, leading
to creating a new one after LIPO linking when we invoke similar handling in
symtab_remove_unreachable_nodes, which then does not have a resolved node. It
makes the handling of such nodes equivalent to that in
symtab_remove_unreachable_nodes.

Tested with regression tests and internal benchmarks. Ok for google/4_9

2014-08-25  Teresa Johnson  <tejohnson@google.com>

	Google ref b/16731481
	* cgraphunit.c (analyze_functions): Don't remove origin node.

From-SVN: r214426