libgomp plugins: Remove the (OpenACC) avail function.
This idea has been to use this to see whether a class of devices is generally
available, without determining the exact number of devices that are available.
Such a distinction probably doesn't make too much sense; in the following, we'd
like to use one device anyway.

Inline the few users.

	libgomp/
	* target.h (struct ACC_dispatch_t): Remove avail_func.  Update all
	users.
	* oacc-host.c (openacc_avail): Likewise.
	* plugin-nvptx.c (PTX_avail, openacc_avail): Likewise.
	* libgomp.texi (acc_get_num_devices): Revise.

From-SVN: r217141