More minor changes for -ftwo-level-line-tables.
With this patch, we output an empty name for non-inlined functions,
and change the -ftwo-level-all-subprogs option to override this
behavior to output linkage names instead. The consumer can obtain
the linkage name for non-inlined functions from the ELF symbol table,
so we can save quite a bit of space in the line tables by omitting
those names.

This patch also fixes a case where we output the linkage name for
a clone rather than for the ultimate origin.

This patch is for the google/gcc-4_9 branch.

2015-03-03  Cary Coutant  <ccoutant@google.com>

	* common.opt (ftwo-level-all-subprogs): Default to off;
        update help text.
	* dwarf2out.c (add_subprog_entry): Clear subprog_num if the
        subprogram is inlined, but was already output as non-inlined.
	(out_subprog_directive): Get name from decl_ultimate_origin.
        Output empty name for non-inlined subprograms.
	(out_logical_entry): Output subprog entries for all subprograms.

From-SVN: r221163