Remove this line:
  { return *static_cast<_Tp*>(nullptr); }
by replacing _S_ref with _S_ptr, so that calling begin() on a zero-sized
std::array doesn't immediately perform indirection on a null pointer.

This also requires marking std::__addressof as constexpr in C++11 mode.
Note that this does not affect std::addressof, as the latter is a wrapper
rather than an alias.

In addition, two tests needed line numbers changed.

From-SVN: r222075