Don't use fnmatch in libgomp.
On reviewing in more detail the changes to pass offloading targets
from the driver to libgomp at link time to identify the minimal
self-contained pieces that can go to trunk, I found that the use of
fnmatch to match against target names was completely unnecessary; the
ISO C90 functions strstr and strncmp could be used instead, so
avoiding the need to add configure tests for fnmatch.  This patch duly
removes the use of and configure tests for fnmatch.

	* plugin/configfrag.ac: Don't test for fnmatch.h or fnmatch.
	* configure, config.h.in: Regenerate.
	* target.c [PLUGIN_SUPPORT]: Don't include <fnmatch.h>.
	(offload_target_to_plugin_name): Use strstr and strncmp instead of
	fnmatch.

From-SVN: r227175