Add support for the profile-func-internal-id parameter to the coverage callback.
2015-09-15  Matt Deeds  <mattdeeds@google.com>

Add support for the profile-func-internal-id parameter to the coverage callback.
Without this change, the function identifier passed to __coverage_callback
(enabled with param=coverage-callback=1) does not match the values emitted in
the .gcno file.  Because the function profile_id is typically more unique
(typically 32 bits) than the function internal id (typically 16 bits), it can be
desirable to have the profile_id used to identify a function as opposed to the
function internal id.
	* gcc/tree-profile.c (gimple_gen_edge_profiler): Add support for the
profile-func-internal-id parameter to the coverage callback

From-SVN: r227810