Fix compile time error introduced by PR64111
This fixes a severe compile time issue that was introduced in GCC
FSF by the initial fix for PR6411.  It came into the ChromeOS
compiler as part of svn r220383, which was a merge from Google 4.9.
It came into Google 4.9 as part of svn r219106, which was a merge
of several revisions from the GCC FSF compiler.

From-SVN: r232625