Use df_read_modify_subreg_p in cprop.c
local_cprop_find_used_regs punted on all multiword registers,
with the comment:

      /* Setting a subreg of a register larger than word_mode leaves
         the non-written words unchanged.  */

But this only applies if the outer mode is smaller than the
inner mode.  If they're the same size then writes to the subreg
are a normal full update.

This patch uses df_read_modify_subreg_p instead.  A later patch
adds more uses of the same routine, but this part had a (positive)
effect on code generation for the testsuite whereas the others
seemed to be simple clean-ups.

From-SVN: r242095