Add a scalar_mode class
This patch adds a scalar_mode class that can hold any scalar mode,
specifically:

  - scalar integers
  - scalar floating-point values
  - scalar fractional modes
  - scalar accumulator modes
  - pointer bounds modes

The type is useful in four main situations:

  - as the mode of a CONST_SCALAR_INT_P.  Logically these would
    be scalar_int_modes, but we also use CONST_INT for pointer
    bounds modes.

  - while operating on the fixed-point modes (fractional and
    accumulator).  It would be possible to define special classes
    for these, but at the moment nothing would benefit from them.
    In particular, there's no use case that would help select between
    one class for all fixed-point modes versus one class for fractional
    modes and one class for accumulator modes.

  - in hooks like scalar_mode_supported_p.

  - as the mode of a vector element, in cases where we don't need
    to distinguish between integer and floating-point elements.

The number of uses is much smaller than for scalar_int_mode,
so the patch handles them all at once.

The patch also gets rid of a redundant assignment to total_bytes
in native_interpret_real.  The variable was already set to the
same value in the variable declaration block.

From-SVN: r242160