Make more use of paradoxical_subreg_p
This patch makes more use of the existing paradoxical_subreg_p
predicate and also adds a version that operates on outer and
inner modes.

Some of the affected tests were based on GET_MODE_SIZE rather than
GET_MODE_PRECISION and so could change the result for modes that
have the same size but different precisions.  In each case the
change should be a no-op or an improvement.  (Specifically, the
big-endian byte offset correction will be 0 for paradoxical subregs
whose byte sizes are the same.)

The patch changes the branch taken for full subregs in
simplify_subreg, but the new form matches the commentary more
closely.  Both branches should be equally good in that situation.

The patch also removes a redundant "GET_CODE (x) == SUBREG"
check from move_plus_up.  By this point we've already checked
that it's a SUBREG and extracted the SUBREG_REG.

From-SVN: r242168