Turn HARD_REGNO_CALL_PART_CLOBBERED into a target hook
Later patches will change the size of a machine_mode from
a compile-time constant to a runtime invariant.  Static
type checking forces target-independent code to treat the
modes as runtime invariants unless it has done the work to
prove that they're not.  However, target-specific code can
continue to treat the modes as constant-sized, if that
assumption is correct for the target.

The main snag with this approach is that target-independent
code still uses target-specific macros.  The best fix for
this would of course be to complete the transtion to using
target hooks instead.  However, that's a lot of work,
especially since some of the macros are still used in
preprocessor conditions.

This patch instead adds a MACRO_MODE wrapper that can be
used around mode arguments to target-specific macros.

The first use case is in a partial conversion of
HARD_REGNO_CALL_PART_CLOBBERED to a target hook:
target-independent code now uses the hook, but ports still
define the macro.  It would be good to convert the ports to
define the hook directly, but that's left as future work.

From-SVN: r242185