Pass rtx and index to read-md.c iterator routines
The read-md.c iterator callbacks previously used a void *
to record the position at which the iterator value should be
installed.  This doesn't scale easily to the SUBREG_BYTE
representation used by a later patch, so this patch replaces
the void * with both an rtx and an operand number.  The operand
number is ignored for modes and codes.

From-SVN: r242197