Pass number of elements alongside tree* when constructing vectors
Several routines take an element array as a "tree *".  This patch
also makes them take the number of elements in the array.

The reason this is useful is that, after later patches, the number
of elements in a general vector can be variable, so there was a
choice between having the affected routines assert that the
number is constant or have the callers pass the constant number
of units down alongside the pointer.  The latter seemed better
since it ensures that the caller and callee are in sync.
Using a vec might in some ways be better, but several callers
want to pass an array splice.

From-SVN: r242202