Let the target choose a vectorisation alignment
The vectoriser previously aligned vectors to TYPE_ALIGN
unconditionally, although there was also a hard-coded
assumption that this was equal to the type size.  This was
inconvenient for SVE for two reasons:

- When compiling for a specific power-of-2 SVE vector length,
  we generally want to align to a full vector.  However, the
  TYPE_ALIGN is governed by the ABI alignment, which is 16 bytes
  regardless of size.

- For vector-length-agnostic code it doesn't make sense to try
  to align, since the runtime vector length might not be a
  power of two.  Even for power of two sizes, there's no
  guarantee that aligning to the previous 16 bytes will be
  an improveent.

This patch therefore adds a target hook to control the preferred
vectoriser (as opposed to ABI) alignment.

From-SVN: r242208