Add VEC_SERIES_EXPR and associated optab
This patch adds a tree equivalent of the VEC_SERIES rtl code that
was added by the previous patch.  It also adds an associated optab.

Like VEC_DUPLICATE_EXPR, the code is treated as a GIMPLE_SINGLE_RHS
rather than a GIMPLE_BINARY_RHS.  This is so that we can treat
constant vector series as gimple values, like a VECTOR_CST would be.

From-SVN: r242218